feat : category repository initialize
카테고리 repository 인터페이스 구현

+++ src/main/java/com/example/demo/src/record/model/PostRecordReq.java
@@ -15,7 +15,7 @@ public class PostRecordReq {
     Long goalIdx;
     String date;
     boolean type;
-    int category;
+    Long category;
     int amount;
 
     public RecordEntity toEntity(UserEntity user, GoalEntity goal, CategoryEntity category) {
